task-2/requirements.txt
apache-airflow==2.2.5
apache-airflow-providers-google==6.0.0
google-api-core==2.8.2
gspread==5.7.1
oauth2client==4.1.3
pandas==1.3.0
pandas-gbq==0.14.1
pymysql==1.0.0
psycopg2-binary
python-dotenv==0.21.0